backend/requirements.txt
fastapi
uvicorn
python-multipart
pdfplumber
groq
openai
python-docx
reportlab
python-dotenv
chromadb
sentence-transformers
